+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- C++ unit tests now include PNG and WebP images in at least a basic fashion
+
 ### Fixed
 
-- Removed the debug printing to `stderr`
+- Removed the debug printing to `stderr` for release builds
 - Fixed an issue where images with no `Xmp.mwg-rs.Regions` key would still try to parse regions
 
 ## [0.2.0] - 2025-06-12